requirements.txt
mysql-connector-python>=9.0.0
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

